refactor: auth 패키지 생성 및 관련 클래스 이동
