+++ src/main/java/nextstep/auth/application/TokenResponse.java
@@ -1,4 +1,4 @@
-package nextstep.member.application;
+package nextstep.auth.application;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
