+++ src/main/java/nextstep/auth/infrastructure/GithubProfileResponse.java
@@ -1,4 +1,4 @@
-package nextstep.member.infrastructure;
+package nextstep.auth.infrastructure;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -7,6 +7,6 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @AllArgsConstructor
 @Getter
-public class GithubAccessTokenResponse {
-    private String accessToken;
+public class GithubProfileResponse {
+    private String email;
 }
