+++ src/main/java/nextstep/auth/presentation/AuthenticationPrincipalArgumentResolver.java
@@ -1,9 +1,9 @@
-package nextstep.member.presentation;
+package nextstep.auth.presentation;
 
 import lombok.RequiredArgsConstructor;
-import nextstep.member.AuthenticationException;
-import nextstep.member.application.JwtTokenProvider;
-import nextstep.member.domain.LoginMember;
+import nextstep.auth.AuthenticationException;
+import nextstep.auth.application.JwtTokenProvider;
+import nextstep.auth.domain.LoginMember;
 import org.springframework.core.MethodParameter;
 import org.springframework.web.bind.support.WebDataBinderFactory;
 import org.springframework.web.context.request.NativeWebRequest;
