+++ src/main/java/nextstep/auth/presentation/ExternalTokenRequest.java
@@ -1,0 +1,10 @@
+package nextstep.auth.presentation;
+
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+
+@RequiredArgsConstructor
+@Getter
+public class ExternalTokenRequest {
+    private final String code;
+}
